Message class
This patch adds a new Message class, which is a Lua wrapper around a
zmq_msg_t instance.  It has a single method, data(), which returns the
content of the message as a Lua string.

In addition, the Socket class now has “raw” versions of send and recv,
which take in Message instances instead of Lua strings.  This lets you
read a message from one socket and write that same message to another
without having to copy the data.